+++ src/main/java/br/ufc/quixada/arquitetura/gvp/servico/IVendaServico.java
@@ -9,7 +9,7 @@ public interface IVendaServico {
 
 	public abstract List<Venda> listar();
 
-	public Venda buscarPorId(Long id);
+	public Venda buscarPorId(Integer id);
 	
 	public void apagar(Venda Venda);
 }
